tests: Separate test suite for dataset questions
