+++ tests/unit/api/v1/test_datasets.py
@@ -14,11 +14,13 @@
 import math
 import uuid
 from datetime import datetime
-from typing import TYPE_CHECKING, Any, Dict, List, Tuple, Type, Union
+from typing import Any, Dict, List, TYPE_CHECKING, Tuple, Type, Union
 from unittest.mock import ANY, MagicMock
 from uuid import UUID, uuid4
 
 import pytest
+from sqlalchemy import func, inspect, select
+
 from argilla_server.apis.v1.handlers.datasets.records import LIST_DATASET_RECORDS_LIMIT_DEFAULT
 from argilla_server.constants import API_KEY_HEADER_NAME
 from argilla_server.enums import (
@@ -51,17 +53,6 @@ from argilla_server.schemas.v1.metadata_properties import (
     METADATA_PROPERTY_CREATE_TITLE_MAX_LENGTH,
     TERMS_METADATA_PROPERTY_VALUES_MAX_ITEMS,
 )
-from argilla_server.schemas.v1.questions import (
-    QUESTION_CREATE_DESCRIPTION_MAX_LENGTH,
-    QUESTION_CREATE_NAME_MAX_LENGTH,
-    QUESTION_CREATE_TITLE_MAX_LENGTH,
-    RANKING_OPTIONS_MAX_ITEMS,
-    RATING_OPTIONS_MAX_ITEMS,
-    RATING_OPTIONS_MIN_ITEMS,
-    VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH,
-    VALUE_TEXT_OPTION_TEXT_MAX_LENGTH,
-    VALUE_TEXT_OPTION_VALUE_MAX_LENGTH,
-)
 from argilla_server.schemas.v1.records import RECORDS_CREATE_MAX_ITEMS, RECORDS_CREATE_MIN_ITEMS
 from argilla_server.schemas.v1.vector_settings import (
     VECTOR_SETTINGS_CREATE_NAME_MAX_LENGTH,
@@ -79,8 +70,6 @@ from argilla_server.search_engine import (
     TextQuery,
     UserResponseStatusFilter,
 )
-from sqlalchemy import func, inspect, select
-
 from tests.factories import (
     AdminFactory,
     AnnotatorFactory,
@@ -1149,506 +1138,6 @@ class TestSuiteDatasets:
 
         assert response.status_code == 404
         assert (await db.execute(select(func.count(Field.id)))).scalar() == 0
-
-    @pytest.mark.parametrize(
-        ("settings", "expected_settings"),
-        [
-            ({"type": "text"}, {"type": "text", "use_markdown": False}),
-            ({"type": "text", "use_markdown": True}, {"type": "text", "use_markdown": True}),
-            ({"type": "text", "use_markdown": False}, {"type": "text", "use_markdown": False}),
-            (
-                {
-                    "type": "rating",
-                    "options": [{"value": 1}, {"value": 2}, {"value": 3}, {"value": 4}, {"value": 5}],
-                },
-                {
-                    "type": "rating",
-                    "options": [{"value": 1}, {"value": 2}, {"value": 3}, {"value": 4}, {"value": 5}],
-                },
-            ),
-            (
-                {
-                    "type": "label_selection",
-                    "options": [
-                        {"value": "positive", "text": "Positive", "description": "Text with a positive sentiment"},
-                        {"value": "negative", "text": "Negative", "description": "Text with a negative sentiment"},
-                        {"value": "neutral", "text": "Neutral", "description": "Text with a neutral sentiment"},
-                    ],
-                },
-                {
-                    "type": "label_selection",
-                    "options": [
-                        {"value": "positive", "text": "Positive", "description": "Text with a positive sentiment"},
-                        {"value": "negative", "text": "Negative", "description": "Text with a negative sentiment"},
-                        {"value": "neutral", "text": "Neutral", "description": "Text with a neutral sentiment"},
-                    ],
-                    "visible_options": None,
-                },
-            ),
-            (
-                {
-                    "type": "label_selection",
-                    "options": [
-                        {"value": "positive", "text": "Positive"},
-                        {"value": "negative", "text": "Negative"},
-                        {"value": "neutral", "text": "Neutral"},
-                    ],
-                    "visible_options": 3,
-                },
-                {
-                    "type": "label_selection",
-                    "options": [
-                        {"value": "positive", "text": "Positive", "description": None},
-                        {"value": "negative", "text": "Negative", "description": None},
-                        {"value": "neutral", "text": "Neutral", "description": None},
-                    ],
-                    "visible_options": 3,
-                },
-            ),
-            (
-                {
-                    "type": "ranking",
-                    "options": [
-                        {"value": "completion-a", "text": "Completion A", "description": "Completion A is the best"},
-                        {"value": "completion-b", "text": "Completion B", "description": "Completion B is the best"},
-                        {"value": "completion-c", "text": "Completion C", "description": "Completion C is the best"},
-                        {"value": "completion-d", "text": "Completion D", "description": "Completion D is the best"},
-                    ],
-                },
-                {
-                    "type": "ranking",
-                    "options": [
-                        {"value": "completion-a", "text": "Completion A", "description": "Completion A is the best"},
-                        {"value": "completion-b", "text": "Completion B", "description": "Completion B is the best"},
-                        {"value": "completion-c", "text": "Completion C", "description": "Completion C is the best"},
-                        {"value": "completion-d", "text": "Completion D", "description": "Completion D is the best"},
-                    ],
-                },
-            ),
-            (
-                {
-                    "type": "ranking",
-                    "options": [
-                        {"value": "completion-a", "text": "Completion A", "description": None},
-                        {"value": "completion-b", "text": "Completion b", "description": None},
-                        {"value": "completion-c", "text": "Completion C", "description": None},
-                        {"value": "completion-d", "text": "Completion D", "description": None},
-                    ],
-                },
-                {
-                    "type": "ranking",
-                    "options": [
-                        {"value": "completion-a", "text": "Completion A", "description": None},
-                        {"value": "completion-b", "text": "Completion b", "description": None},
-                        {"value": "completion-c", "text": "Completion C", "description": None},
-                        {"value": "completion-d", "text": "Completion D", "description": None},
-                    ],
-                },
-            ),
-        ],
-    )
-    async def test_create_dataset_question(
-        self,
-        async_client: "AsyncClient",
-        db: "AsyncSession",
-        owner_auth_header: dict,
-        settings: dict,
-        expected_settings: dict,
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": settings,
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 201
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 1
-
-        response_body = response.json()
-        assert await db.get(Question, UUID(response_body["id"]))
-        assert response_body == {
-            "id": str(UUID(response_body["id"])),
-            "name": "name",
-            "title": "title",
-            "description": None,
-            "required": False,
-            "settings": expected_settings,
-            "dataset_id": str(UUID(response_body["dataset_id"])),
-            "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
-            "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
-        }
-
-    async def test_create_dataset_question_with_description(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "description": "description",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 201
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 1
-
-        response_body = response.json()
-        assert await db.get(Question, UUID(response_body["id"]))
-        assert response_body["description"] == "description"
-
-    async def test_create_dataset_question_without_authentication(
-        self, async_client: "AsyncClient", db: "AsyncSession"
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(f"/api/v1/datasets/{dataset.id}/questions", json=question_json)
-
-        assert response.status_code == 401
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_as_admin(self, async_client: "AsyncClient", db: "AsyncSession"):
-        workspace = await WorkspaceFactory.create()
-        admin = await AdminFactory.create(workspaces=[workspace])
-        dataset = await DatasetFactory.create(workspace=workspace)
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions",
-            headers={API_KEY_HEADER_NAME: admin.api_key},
-            json=question_json,
-        )
-
-        assert response.status_code == 201
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 1
-
-    async def test_create_dataset_question_as_admin_for_different_workspace(
-        self, async_client: "AsyncClient", db: "AsyncSession"
-    ):
-        workspace = await WorkspaceFactory.create()
-        admin = await AdminFactory.create(workspaces=[workspace])
-
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions",
-            headers={API_KEY_HEADER_NAME: admin.api_key},
-            json=question_json,
-        )
-
-        assert response.status_code == 403
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_as_annotator(self, async_client: "AsyncClient", db: "AsyncSession"):
-        annotator = await AnnotatorFactory.create()
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions",
-            headers={API_KEY_HEADER_NAME: annotator.api_key},
-            json=question_json,
-        )
-
-        assert response.status_code == 403
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    @pytest.mark.parametrize("invalid_name", ["", " ", "  ", "-", "--", "_", "__", "A", "AA", "invalid_nAmE"])
-    async def test_create_dataset_question_with_invalid_name(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict, invalid_name: str
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": invalid_name,
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_with_invalid_max_length_name(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "a" * (QUESTION_CREATE_NAME_MAX_LENGTH + 1),
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_with_invalid_max_length_title(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "a" * (QUESTION_CREATE_TITLE_MAX_LENGTH + 1),
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_with_invalid_max_length_description(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "description": "a" * (QUESTION_CREATE_DESCRIPTION_MAX_LENGTH + 1),
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_with_existent_name(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        question = await QuestionFactory.create(name="name")
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{question.dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 409
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 1
-
-    async def test_create_dataset_question_with_published_dataset(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        dataset = await DatasetFactory.create(status=DatasetStatus.ready)
-        question_json = {
-            "name": "name",
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert response.json() == {"detail": "Question cannot be created for a published dataset"}
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    async def test_create_dataset_question_with_nonexistent_dataset_id(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
-    ):
-        await DatasetFactory.create()
-        question_json = {
-            "name": "text",
-            "title": "Text",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{uuid4()}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 404
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
-    @pytest.mark.parametrize(
-        "settings",
-        [
-            None,
-            {},
-            {"type": "wrong"},
-            {"type": "text", "use_markdown": None},
-            {"type": "text", "use_markdown": "wrong"},
-            {
-                "type": "rating",
-                "options": [
-                    {"value": "A wrong value"},
-                    {"value": "B wrong value"},
-                    {"value": "C wrong value"},
-                    {"value": "D wrong value"},
-                ],
-            },
-            {"type": "rating", "options": [{"value": value} for value in range(0, RATING_OPTIONS_MIN_ITEMS - 1)]},
-            {"type": "rating", "options": [{"value": value} for value in range(0, RATING_OPTIONS_MAX_ITEMS + 1)]},
-            {"type": "rating", "options": "invalid"},
-            {"type": "rating", "options": [{"value": 1}, {"value": 1}]},
-            {"type": "rating", "options": [{"value": 1}, {"value": 2}, {"value": 13}]},
-            {"type": "rating", "options": [{"value": 1}, {"value": 2}, {"value": -13}]},
-            {"type": "label_selection", "options": []},
-            {"type": "label_selection", "options": [{"value": "just_one_label", "text": "Just one label"}]},
-            {
-                "type": "label_selection",
-                "options": [{"value": "a", "text": "a"}, {"value": "b", "text": "b"}],
-                "visible_options": 0,
-            },
-            {
-                "type": "label_selection",
-                "options": [{"value": "a", "text": "a"}, {"value": "b", "text": "b"}],
-                "visible_options": -1,
-            },
-            {
-                "type": "label_selection",
-                "options": [{"value": "", "text": "a"}, {"value": "b", "text": "b"}],
-            },
-            {
-                "type": "label_selection",
-                "options": [
-                    {"value": "".join(["a" for _ in range(VALUE_TEXT_OPTION_VALUE_MAX_LENGTH + 1)]), "text": "a"},
-                    {"value": "b", "text": "b"},
-                ],
-            },
-            {
-                "type": "label_selection",
-                "options": [{"value": "a", "text": ""}, {"value": "b", "text": "b"}],
-            },
-            {
-                "type": "label_selection",
-                "options": [
-                    {"value": "a", "text": "".join(["a" for _ in range(VALUE_TEXT_OPTION_TEXT_MAX_LENGTH + 1)])},
-                    {"value": "b", "text": "b"},
-                ],
-            },
-            {
-                "type": "label_selection",
-                "options": [{"value": "a", "text": "a", "description": ""}, {"value": "b", "text": "b"}],
-            },
-            {
-                "type": "label_selection",
-                "options": [
-                    {
-                        "value": "a",
-                        "text": "a",
-                        "description": "".join(["a" for _ in range(VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH + 1)]),
-                    },
-                    {"value": "b", "text": "b"},
-                ],
-            },
-            {
-                "type": "label_selection",
-                "options": [
-                    {"value": "a", "text": "a", "description": "a"},
-                    {"value": "b", "text": "b", "description": "b"},
-                    {"value": "b", "text": "b", "description": "b"},
-                ],
-            },
-            {
-                "type": "label_selection",
-                "options": [
-                    {"value": "a", "text": "a", "description": "a"},
-                    {"value": "b", "text": "b", "description": "b"},
-                    {"value": "b", "text": "b", "description": "b"},
-                ],
-                "visible_options": 2,
-            },
-            {
-                "type": "label_selection",
-                "options": [
-                    {"value": "a", "text": "a", "description": "a"},
-                    {"value": "b", "text": "b", "description": "b"},
-                    {"value": "b", "text": "b", "description": "b"},
-                ],
-                "visible_options": 5,
-            },
-            {
-                "type": "ranking",
-                "options": [
-                    {"value": "a", "text": "a", "description": "a"},
-                ],
-            },
-            {
-                "type": "ranking",
-                "options": [
-                    {"value": "a", "text": "a", "description": "a"},
-                    {"value": "b", "text": "b", "description": "b"},
-                    {"value": "b", "text": "b", "description": "b"},
-                ],
-            },
-            {
-                "type": "ranking",
-                "options": [
-                    {"value": value, "text": value, "description": value}
-                    for value in range(0, RANKING_OPTIONS_MAX_ITEMS + 1)
-                ],
-            },
-            {
-                "type": "ranking",
-                "options": [
-                    {
-                        "value": "a",
-                        "text": "a",
-                        "description": "".join(["a" for _ in range(VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH + 1)]),
-                    },
-                    {"value": "b", "text": "b", "description": "b"},
-                ],
-            },
-        ],
-    )
-    async def test_create_dataset_question_with_invalid_settings(
-        self,
-        async_client: "AsyncClient",
-        db: "AsyncSession",
-        owner_auth_header: dict,
-        settings: dict,
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {"name": "question", "title": "Question", "settings": settings}
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
 
     @pytest.mark.parametrize(
         ("settings", "expected_settings"),
